Camera Fixed! Unfortunately...
Actually, this time. The issue was with the code to keep the player on the ground(oddly enough). Gravity for the players rigidbody was on and in order to prevent it from falling though the ground someone put a line in that counteracted gravity with the characterController. But because the position isn't exact, the players transform was jumping between the correct position, and increasingly lower positions. For some reason, the visual wasn't moving, but it was affecting the vector between the player and the enemy, which the camera's rotation was dependent on, which the players rotation was dependent on.
So, I revamped the motion code in player so that the motion is independent of camera, and I deleted the line that was counteracting gravity.
Unfortunately, this meant I had to turn off gravity to prevent the thing from falling through the floor, and I had to temporarily disable jump. I'm gonna start working on getting those fixed.
I also added a camera smoother to the camera base so that the motion of the camera motion could be smoothed a little more beyond the seizure fix.