Added Input->Action Structure
Added classes InputAccessor and ActionManager, and abstract class ActionExecuter. Also added the test child class ChildAE for testing.

The InputAccessor can be used to manage inputs easily for any character. An additional script must be created to actually control the inputs but the InputAccessor regulates the inputs so that they can be easily accessed by any ActionManager and ActionExecuter. The ActionExecuter simply takes an InputAccessor and a child of ActionExecuter as parameters. It then uses IEnumerators to control the wait time between calls to the ActionExecutor so that no more than on action can be executed at once. It passes the InputAccessor and the possibility for combos to the ActionExecuter which passes back an action time and a post action combo time. The ActionManager then wait these amounts of time before calling doAction again.

The idea is that this structure can be easily used for all characters by creating a script to register inputs (either from a keyboard or an AI) and a script to execute actions (attacks, moves blocks, charges, spells, etc).

Later, the ActionManager can be used to store inputs for prekeying (like in smash ultimate).