Ability Library added
The ability library takes all the ability stuff and puts it in its own script for cleanliness. Also all cooldown stuff is dealt with by the new Ability class.
I also added the start for a new ability: fire spell. We don't have to keep but it's mostly to exemplify the ease of use of the AbilityLibrary.
I have more stuff I'll add to ability library later.