Finished the camera changes!
I finished making all the camera changes! There's now two camera views- one lock-on view that looks at the enemy over the player's shoulder, and one free cam view that lets the player look around with the mouse. There's also a targeting system that lets the player choose who to target; while in free cam mode, whichever destructible object is closest is targeted. In order to add all of this to the scene, you just need to add the player prefab, the freelookcamera prefab, and the main camera prefab. Add all of those to the scene and it'll work.